+++ packages/contracts-bedrock/contracts/universal/ProxyAdmin.sol
@@ -6,25 +6,39 @@ import { Owned } from "@rari-capital/solmate/src/auth/Owned.sol";
 import { Lib_AddressManager } from "../legacy/Lib_AddressManager.sol";
 import { L1ChugSplashProxy } from "../legacy/L1ChugSplashProxy.sol";
 
+// Define static interfaces of these proxies so that we can easily
+// use staticcall on the getters we need.
+interface IStatic_ERC1967Proxy {
+    function implementation() external view returns (address);
+
+    function admin() external view returns (address);
+}
+
+interface IStatic_L1ChugSplashProxy {
+    function getImplementation() external view returns (address);
+
+    function getOwner() external view returns (address);
+}
+
 /**
  * @title ProxyAdmin
- * @dev This is an auxiliary contract meant to be assigned as the admin of a Proxy, based on
- *      the OpenZeppelin implementation. It has backwards compatibility logic to work with the
- *      various types of proxies that have been deployed by Optimism.
+ * @dev This is an auxiliary contract meant to be assigned as the admin of an ERC1967 Proxy,
+ *      based on the OpenZeppelin implementation. It has backwards compatibility logic to work with
+ *      the various types of proxies that have been deployed by Optimism.
  */
 contract ProxyAdmin is Owned {
     /**
      * @notice The proxy types that the ProxyAdmin can manage.
      *
-     * @custom:value OpenZeppelin     Represents the OpenZeppelin style transparent proxy
-     *                                interface, this is the standard.
+     * @custom:value ERC1967          Represents an ERC1967 compliant transparent proxy
+     *                                interface, this is the default.
      * @custom:value Chugsplash       Represents the Chugsplash proxy interface,
      *                                this is legacy.
      * @custom:value ResolvedDelegate Represents the ResolvedDelegate proxy
      *                                interface, this is legacy.
      */
     enum ProxyType {
-        OpenZeppelin,
+        ERC1967,
         Chugsplash,
         ResolvedDelegate
     }
@@ -88,8 +102,8 @@ contract ProxyAdmin is Owned {
      *
      * @param _address The address of the address manager.
      */
-    function setAddressManager(address _address) external onlyOwner {
-        addressManager = Lib_AddressManager(_address);
+    function setAddressManager(Lib_AddressManager _address) external onlyOwner {
+        addressManager = _address;
     }
 
     /**
@@ -135,28 +149,15 @@ contract ProxyAdmin is Owned {
     function getProxyImplementation(Proxy proxy) external view returns (address) {
         ProxyType proxyType = proxyType[address(proxy)];
 
-        // We need to manually run the static call since the getter cannot be flagged as view
-        address target;
-        bytes memory data;
-        if (proxyType == ProxyType.OpenZeppelin) {
-            target = address(proxy);
-            data = abi.encodeWithSelector(Proxy.implementation.selector);
+        if (proxyType == ProxyType.ERC1967) {
+            return IStatic_ERC1967Proxy(address(proxy)).implementation();
         } else if (proxyType == ProxyType.Chugsplash) {
-            target = address(proxy);
-            data = abi.encodeWithSelector(L1ChugSplashProxy.getImplementation.selector);
+            return IStatic_L1ChugSplashProxy(address(proxy)).getImplementation();
         } else if (proxyType == ProxyType.ResolvedDelegate) {
-            target = address(addressManager);
-            data = abi.encodeWithSelector(
-                Lib_AddressManager.getAddress.selector,
-                implementationName[address(proxy)]
-            );
+            return addressManager.getAddress(implementationName[address(proxy)]);
         } else {
             revert("ProxyAdmin: unknown proxy type");
         }
-
-        (bool success, bytes memory returndata) = target.staticcall(data);
-        require(success);
-        return abi.decode(returndata, (address));
     }
 
     /**
@@ -169,25 +170,15 @@ contract ProxyAdmin is Owned {
     function getProxyAdmin(Proxy proxy) external view returns (address) {
         ProxyType proxyType = proxyType[address(proxy)];
 
-        // We need to manually run the static call since the getter cannot be flagged as view
-        address target;
-        bytes memory data;
-        if (proxyType == ProxyType.OpenZeppelin) {
-            target = address(proxy);
-            data = abi.encodeWithSelector(Proxy.admin.selector);
+        if (proxyType == ProxyType.ERC1967) {
+            return IStatic_ERC1967Proxy(address(proxy)).admin();
         } else if (proxyType == ProxyType.Chugsplash) {
-            target = address(proxy);
-            data = abi.encodeWithSelector(L1ChugSplashProxy.getOwner.selector);
+            return IStatic_L1ChugSplashProxy(address(proxy)).getOwner();
         } else if (proxyType == ProxyType.ResolvedDelegate) {
-            target = address(addressManager);
-            data = abi.encodeWithSignature("owner()");
+            return addressManager.owner();
         } else {
             revert("ProxyAdmin: unknown proxy type");
         }
-
-        (bool success, bytes memory returndata) = target.staticcall(data);
-        require(success);
-        return abi.decode(returndata, (address));
     }
 
     /**
@@ -200,12 +191,14 @@ contract ProxyAdmin is Owned {
     function changeProxyAdmin(Proxy proxy, address newAdmin) external onlyOwner {
         ProxyType proxyType = proxyType[address(proxy)];
 
-        if (proxyType == ProxyType.OpenZeppelin) {
+        if (proxyType == ProxyType.ERC1967) {
             proxy.changeAdmin(newAdmin);
         } else if (proxyType == ProxyType.Chugsplash) {
             L1ChugSplashProxy(payable(proxy)).setOwner(newAdmin);
         } else if (proxyType == ProxyType.ResolvedDelegate) {
-            Lib_AddressManager(addressManager).transferOwnership(newAdmin);
+            addressManager.transferOwnership(newAdmin);
+        } else {
+            revert("ProxyAdmin: unknown proxy type");
         }
     }
 
@@ -218,7 +211,7 @@ contract ProxyAdmin is Owned {
     function upgrade(Proxy proxy, address implementation) public onlyOwner {
         ProxyType proxyType = proxyType[address(proxy)];
 
-        if (proxyType == ProxyType.OpenZeppelin) {
+        if (proxyType == ProxyType.ERC1967) {
             proxy.upgradeTo(implementation);
         } else if (proxyType == ProxyType.Chugsplash) {
             L1ChugSplashProxy(payable(proxy)).setStorage(
@@ -227,7 +220,9 @@ contract ProxyAdmin is Owned {
             );
         } else if (proxyType == ProxyType.ResolvedDelegate) {
             string memory name = implementationName[address(proxy)];
-            Lib_AddressManager(addressManager).setAddress(name, implementation);
+            addressManager.setAddress(name, implementation);
+        } else {
+            revert("ProxyAdmin: unknown proxy type");
         }
     }
 
@@ -246,9 +241,10 @@ contract ProxyAdmin is Owned {
     ) external payable onlyOwner {
         ProxyType proxyType = proxyType[address(proxy)];
 
-        if (proxyType == ProxyType.OpenZeppelin) {
+        if (proxyType == ProxyType.ERC1967) {
             proxy.upgradeToAndCall{ value: msg.value }(implementation, data);
         } else {
+            // reverts if proxy type is unknown
             upgrade(proxy, implementation);
             (bool success, ) = address(proxy).call{ value: msg.value }(data);
             require(success);