+++ packages/contracts-periphery/contracts/testing/helpers/ExternalContractCompiler.sol
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.9;
+
+import { ProxyAdmin } from "@eth-optimism/contracts-bedrock/contracts/universal/ProxyAdmin.sol";
+import { Proxy } from "@eth-optimism/contracts-bedrock/contracts/universal/Proxy.sol";
+
+/**
+ * Just exists so we can compile external contracts.
+ */
+contract ExternalContractCompiler {
+
+}